+++ internal/js_ast/js_ast.go
@@ -1658,7 +1658,9 @@ func GenerateNonUniqueNameFromPath(path string) string {
 		}
 	}
 
-	// Convert it to an ASCII identifier
+	// Convert it to an ASCII identifier. Note: If you change this to a non-ASCII
+	// identifier, you're going to potentially cause trouble with non-BMP code
+	// points in target environments that don't support bracketed Unicode escapes.
 	bytes := []byte{}
 	needsGap := false
 	for _, c := range base {